+++ storage/secondary.go
@@ -105,47 +105,3 @@ func (s *secondaryQuerier) Select(sortSeries bool, hints *SelectHints, matchers 
 	}
 	return &lazySeriesSet{create: s.createFn(s.genericQuerier.Select(sortSeries, hints, matchers...))}
 }
-
-type lazySeriesSet struct {
-	create func() (s genericSeriesSet, ok bool)
-
-	set genericSeriesSet
-}
-
-func (c *lazySeriesSet) Next() bool {
-	if c.set != nil {
-		return c.set.Next()
-	}
-
-	var ok bool
-	c.set, ok = c.create()
-	return ok
-}
-
-func (c *lazySeriesSet) Err() error {
-	if c.set != nil {
-		return c.set.Err()
-	}
-	return nil
-}
-
-func (c *lazySeriesSet) At() Labels {
-	if c.set != nil {
-		return c.set.At()
-	}
-	return nil
-}
-
-func (c *lazySeriesSet) Warnings() Warnings {
-	if c.set != nil {
-		return c.set.Warnings()
-	}
-	return nil
-}
-
-type warningsOnlySeriesSet Warnings
-
-func (warningsOnlySeriesSet) Next() bool           { return false }
-func (warningsOnlySeriesSet) Err() error           { return nil }
-func (warningsOnlySeriesSet) At() Labels           { return nil }
-func (c warningsOnlySeriesSet) Warnings() Warnings { return Warnings(c) }
